Corrected time.passed
I was not getting any difference in gamma stat despite higher births to
deaths. I think this is because I was up-weighting the amount of time
passed for every add() by calculating the amount of time passed using
birth+death. This means that as I increased births to deaths, the
increase in births was down weighted by an increase in time passed.